+++ src/geos.h
@@ -47,7 +47,8 @@ enum {
     PGERR_NOT_A_GEOMETRY,
     PGERR_GEOS_EXCEPTION,
     PGERR_NO_MALLOC,
-    PGERR_GEOMETRY_TYPE
+    PGERR_GEOMETRY_TYPE,
+    PGERR_MULTIPOINT_WITH_POINT_EMPTY
 };
 
 // Define how the states are handled by CPython
@@ -64,6 +65,8 @@ enum {
         PyErr_SetString(PyExc_MemoryError, "Could not allocate memory"); break;\
       case PGERR_GEOMETRY_TYPE:\
         PyErr_SetString(PyExc_TypeError, "One of the Geometry inputs is of incorrect geometry type."); break;\
+      case PGERR_MULTIPOINT_WITH_POINT_EMPTY:\
+        PyErr_SetString(PyExc_ValueError, "WKT output of multipoints with an empty point is unsupported on this version of GEOS."); break;\
       default:\
         PyErr_Format(PyExc_RuntimeError, "Pygeos ufunc returned with unknown error state code %d.", errstate); break;\
     }
@@ -108,6 +111,7 @@ extern PyObject *geos_exception[1];
 
 extern void geos_error_handler(const char *message, void *userdata);
 extern void geos_notice_handler(const char *message, void *userdata);
+extern char check_to_wkt_compatible(GEOSContextHandle_t ctx, GEOSGeometry *geom);
 
 extern int init_geos(PyObject *m);