+++ CHANGELOG.rst
@@ -6,11 +6,12 @@ Version 0.8 (unreleased)
 
 **Highlights of this release**
 
+* Prevent segfault in to_wkt (and repr) with empty points in multipoints (#171)
 * Fixed bug in ``multilinestrings()``, it now accepts linearrings again (#168)
 * Release the GIL to allow for multithreading in functions that do not
   create geometries (#144) and in the STRtree ``query_bulk()`` method (#174)
 * Addition of a ``frechet_distance()`` function for GEOS >= 3.7 (#144)
-* Addition of ``coverage_union()`` and ``coverage_union_all()` functions
+* Addition of ``coverage_union()`` and ``coverage_union_all()`` functions
   for GEOS >= 3.8 (#142)
 * Fixed segfaults when adding empty geometries to the STRtree (#147)
 * Addition of ``include_z=True`` keyword in the ``get_coordinates()`` function